+++ OmmelSamvirke.Domain/Features/Pages/Interfaces/Repositories/IContentBlockRepository.cs
@@ -1,0 +1,9 @@
+﻿using OmmelSamvirke.Domain.Common.Interfaces;
+using OmmelSamvirke.Domain.Features.Pages.Models.ContentBlocks;
+
+namespace OmmelSamvirke.Domain.Features.Pages.Interfaces.Repositories;
+
+public interface IContentBlockRepository : IGenericRepository<ContentBlock>
+{
+    
+}