+++ OmmelSamvirke.Domain/Features/Pages/Interfaces/Repositories/IPageTemplateRepository.cs
@@ -1,0 +1,9 @@
+﻿using OmmelSamvirke.Domain.Common.Interfaces;
+using OmmelSamvirke.Domain.Features.Pages.Models;
+
+namespace OmmelSamvirke.Domain.Features.Pages.Interfaces.Repositories;
+
+public interface IPageTemplateRepository : IGenericRepository<PageTemplate>
+{
+
+}