+++ OmmelSamvirke.Domain/Features/Pages/Interfaces/Repositories/IPdfBlockDataRepository.cs
@@ -1,0 +1,9 @@
+﻿using OmmelSamvirke.Domain.Common.Interfaces;
+using OmmelSamvirke.Domain.Features.Pages.Models.ContentBlockData;
+
+namespace OmmelSamvirke.Domain.Features.Pages.Interfaces.Repositories;
+
+public interface IPdfBlockDataRepository : IGenericRepository<PdfBlockData>
+{
+    
+}